requirements-base.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
fake-useragent>=1.4.0
pandas>=2.0.0
openpyxl>=3.1.0
aiohttp>=3.8.0
openai>=1.0.0